requirements.txt
Flask==3.0.0
numpy==1.26.4
Pillow==10.4.0
onnxruntime-cpu
rembg==2.0.60
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.41.0